fix(static-analysis): curate clang-tidy into an enforcing gate
From the last green run's ~1100 findings: 441 were misc-include-cleaner
noise from Zephyr umbrella headers, 223 were -Wconversion promotions the
GCC hardening policy already treats as warn-only, ~90 sat in vendored
monocypher (safety-policy exempt, now excluded from the tidy file list),
and six check classes were false-positive-prone for this codebase (each
disabled with rationale in lichen/.clang-tidy).

The remaining real findings are fixed: a duplicated switch branch in
coap_oscore.c, a uint8_t loop index iterating to a size_t bound in
app_interface.c, and three implementation-defined size_t->int narrowings
in bitstream.c made explicit.

The CI step drops continue-on-error. Note: clang-tidy is not installable
locally, so the zero-findings state is validated by the next CI run.

+++ lichen/subsys/lichen/app_interface/app_interface.c
@@ -81,10 +81,10 @@ int lichen_app_interface_register_sink(
 	}
 
 	k_mutex_lock(&s_mutex, K_FOREVER);
-	for (uint8_t i = 0U; i < ARRAY_SIZE(s_sinks); i++) {
+	for (size_t i = 0U; i < ARRAY_SIZE(s_sinks); i++) {
 		if (s_sinks[i].used && same_sink(&s_sinks[i].sink, sink)) {
 			if (out_id != NULL) {
-				*out_id = i;
+				*out_id = (uint8_t)i;
 			}
 			k_mutex_unlock(&s_mutex);
 			return 0;
